create custom bottom sheet component with clicalbe backdrop
